requirements.txt
streamlit
langchain
google-generativeai
langchain_google_genai
python-dotenv